feat: * update quiz models cy-54
