+++ apps/backend/src/modules/quiz/libs/models/models.ts
@@ -1,2 +1,2 @@
-export { QuestionOptionsModel } from "./question-options.model.js";
-export { QuestionModel } from "./questions.model.js";
+export { QuestionOptionModel } from "./question-option.model.js";
+export { QuestionModel } from "./question.model.js";
